requirements.txt
openai>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-go>=0.20.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
gitpython>=3.1.40
